Refactor the kind deployer
