+++ kubetest2-kind/deployer/deployer.go
@@ -20,14 +20,9 @@ package deployer
 import (
 	"os"
 	"path/filepath"
-	"strconv"
-	"strings"
 
 	"github.com/spf13/pflag"
 
-	"sigs.k8s.io/kubetest2/pkg/exec"
-	"sigs.k8s.io/kubetest2/pkg/metadata"
-	"sigs.k8s.io/kubetest2/pkg/process"
 	"sigs.k8s.io/kubetest2/pkg/types"
 )
 
@@ -48,7 +43,6 @@ func New(opts types.Options) (types.Deployer, *pflag.FlagSet) {
 // assert that New implements types.NewDeployer
 var _ types.NewDeployer = New
 
-// TODO(bentheelder): finish implementing this stubbed-out deployer
 type deployer struct {
 	// generic parts
 	commonOptions types.Options
@@ -107,107 +101,6 @@ func bindFlags(d *deployer) *pflag.FlagSet {
 
 // assert that deployer implements types.DeployerWithKubeconfig
 var _ types.DeployerWithKubeconfig = &deployer{}
-
-// Deployer implementation methods below
-
-func (d *deployer) Up() error {
-	args := []string{
-		"create", "cluster",
-		"--name", d.clusterName,
-	}
-	if d.logLevel != "" {
-		args = append(args, "--loglevel", d.logLevel)
-	}
-	// set the explicitly specified image name if set
-	if d.nodeImage != "" {
-		args = append(args, "--image", d.nodeImage)
-	} else if d.commonOptions.ShouldBuild() {
-		// otherwise if we just built an image, use that
-		// NOTE: this is safe in the face of upstream changes, because
-		// we use the same logic / constant for Build()
-		args = append(args, "--image", kindDefaultBuiltImageName)
-	}
-	if d.configPath != "" {
-		args = append(args, "--config", d.configPath)
-	}
-	if d.kubeconfigPath != "" {
-		args = append(args, "--kubeconfig", d.kubeconfigPath)
-	}
-	if d.verbosity > 0 {
-		args = append(args, "--verbosity", strconv.Itoa(d.verbosity))
-	}
-
-	println("Up(): creating kind cluster...\n")
-	// we want to see the output so use process.ExecJUnit
-	return process.ExecJUnit("kind", args, os.Environ())
-}
-
-func (d *deployer) Down() error {
-	args := []string{
-		"delete", "cluster",
-		"--name", d.clusterName,
-	}
-	if d.logLevel != "" {
-		args = append(args, "--loglevel", d.logLevel)
-	}
-
-	println("Down(): deleting kind cluster...\n")
-	// we want to see the output so use process.ExecJUnit
-	return process.ExecJUnit("kind", args, os.Environ())
-}
-
-func (d *deployer) IsUp() (up bool, err error) {
-	// naively assume that if the api server reports nodes, the cluster is up
-	lines, err := exec.CombinedOutputLines(
-		exec.Command("kubectl", "get", "nodes", "-o=name"),
-	)
-	if err != nil {
-		return false, metadata.NewJUnitError(err, strings.Join(lines, "\n"))
-	}
-	return len(lines) > 0, nil
-}
-
-func (d *deployer) DumpClusterLogs() error {
-	args := []string{
-		"export", "logs",
-		"--name", d.clusterName,
-		d.logsDir,
-	}
-	if d.logLevel != "" {
-		args = append(args, "--loglevel", d.logLevel)
-	}
-
-	println("DumpClusterLogs(): exporting kind cluster logs...\n")
-	// we want to see the output so use process.ExecJUnit
-	return process.ExecJUnit("kind", args, os.Environ())
-}
-
-func (d *deployer) Build() error {
-	// TODO(bentheelder): build type should be configurable
-	args := []string{
-		"build", "node-image",
-	}
-	if d.logLevel != "" {
-		args = append(args, "--loglevel", d.logLevel)
-	}
-	if d.buildType != "" {
-		args = append(args, "--type", d.buildType)
-	}
-	if d.kubeRoot != "" {
-		args = append(args, "--kube-root", d.kubeRoot)
-	}
-	// set the explicitly specified image name if set
-	if d.nodeImage != "" {
-		args = append(args, "--image", d.nodeImage)
-	} else if d.commonOptions.ShouldBuild() {
-		// otherwise if we just built an image, use that
-		args = append(args, "--image", kindDefaultBuiltImageName)
-	}
-
-	println("Build(): building kind node image...\n")
-	// we want to see the output so use process.ExecJUnit
-	return process.ExecJUnit("kind", args, os.Environ())
-}
 
 // well-known kind related constants
 const kindDefaultBuiltImageName = "kindest/node:latest"
